# [Optimization][Perf] Disable the GC during CUDA graph capture to speed up by up to 3x